Reapply "Fix incorrect instancing model shadows"
In addition to the original patch, this commit adds two new properties
to set the shadow map bounds of the instance table, namely
shadowBoundsMinimum and shadowBoundsMaximum. They can be used to specify
the shadow map bounds of the instance table. It is used only when max >
min, otherwise the bounds are calculated by going through each instance
and calculating its bounds.

This gives the user the opportunity to specify the bounds if the
performance when calulcating them is too bad.

[ChangeLog][Instancing] Add shadowBoundsMinimum and shadowBoundsMaximum
properties.

Amends 34596dc64548a53626ac910ed947f7751b37da7b and
778e505fbb57231f5943716c906442240ce27b30

Fixes: QTBUG-125876
Change-Id: I23f55274f56f4f13db4df99bb692570f7f456c64
Reviewed-by: Paul Olav Tvete <paul.tvete@qt.io>