OpenXR: Add Metal backend
This is necessary to support running the Meta XR Simulator on macOS. The
Meta XR Simulator provides and XR runtime, but only supports using
Vulkan or Metal. We have a Vulkan backend, but Vulkan on macOS is not
really supported by Qt. So a new Metal backend for OpenXR was added for
macOS, which enables OpenXR applications to be run on macOS when an
OpenXR Runtime is present.

Change-Id: Idfdd238209b9bc0eb0355fe0e27044ee39c99c2d
Reviewed-by: Christian Strømme <christian.stromme@qt.io>