visionOS: Fix dedicate rendering mode
This amends d945aaf9bed2cc5ca1010aa904c84b217bf5dc27 as the dedicated
rendering was adopted to work with threaded rendering. With this change
we re-enable the dedicated mode making it possible not force disabling
multiview and again make it possible to run applications in the
emulator, which does not support multiview rendering.

Pick-to: 6.8
Change-Id: I7f9c62cc891cedae6c6fa75fb451fe46db7c5e35
Reviewed-by: Andy Nichols <andy.nichols@qt.io>