visionOS: Add VRR support
This change enables foveated rendering by using the shading rate maps.
Foveation is on by default on the device and should improve performance
since by reducing the pressure on the pixel shader. Foveation is enabled
both for both the dedicated and layered rendering mode. Note however, as
with multiview, foveated rendering is not supported in the emulator.

Fixes: QTBUG-129733
Pick-to: 6.8
Change-Id: I9e8876ea699baf01abf1b0a9dd0683dff4f2ddc0
Reviewed-by: Andy Nichols <andy.nichols@qt.io>