Xr: Fix embedded View3Ds in XR when rendering with multiview
This adds a fix when View3Ds are used inside a XrView and with mutltiview
enabled. Since the view count was per RHI context, embedded View3Ds,
which are rendered to an offscreen texture, would pick up the same view
count as used for the XrView. To fix this we now store the view count on
the layer, as that is per View3D.

While the solution here is not platform specific, it's unfortunately not
the complete solution for visionOS, where further research will be needed.

Pick-to: 6.9 6.8
Task-number: QTBUG-130962
Change-Id: Iaa603338eba6d4aa0186c53b211615036466466b
Reviewed-by: Paul Olav Tvete <paul.tvete@qt.io>