Fix usage of aggregate type for the extension contexts
The ExtensionContext type, which is masquerading as an aggregate,
is triggering an issue in GCC > 11.4, causing the creation of the
initialization list for the storage type (vector) to be created as an
empty list, when the intention was to have one initial item in the list.
When Qt was built with GCC > 11.4 this would cause the extension system
to not work correctly, as the storage sizes would get out of sync and
the extension contexts would be ignored.
Despite this being a GCC specific issue, the ExtensionContext should
never have been an aggregate, as it's a complex type and as this shows
we get into all the pitfalls of aggregates. To avoid this issue
altogether the ExtensionContext type is now a non-aggregate,
and we explicit create the first and initial type in the
initialization list. Note that while this issue specific for the list of
ExtensionContexts, the other vectors used here also was changed to have
their type set explicitly in the initialization list.

Fixes: QTBUG-132273
Pick-to: 6.9 6.8
Change-Id: I47bc9a921101c3794d68eec9cde4c2eb34c803fc
Reviewed-by: Volker Hilsheimer <volker.hilsheimer@qt.io>